refactoring extras folder
