+++ ontospy/extras/dbpedia.py
