+++ ontospy/extras/hacks/compare.py
