+++ ontospy/extras/hacks/matcher.py
