+++ ontospy/extras/hacks/pptoolkit.py
