+++ ontospy/extras/hacks/rdf-schema.ttl
