+++ ontospy/extras/hacks/sketch.py
