+++ ontospy/extras/hacks/sparqlstore.py
