+++ ontospy/extras/hacks/vocabsturtleprompt.py
