+++ ontospy/extras/shell.py
@@ -7,7 +7,7 @@ ONTOSPY
 Copyright (c) 2013-2017 __Michele Pasin__ <http://www.michelepasin.org>.
 All rights reserved.
 
-More info in the README file.
+OntoSpy Shell Launcher
 
 """
 
@@ -24,7 +24,7 @@ except:
     sys.exit(0)
 
 from .. import *  # imports __init__
-from .shell import Shell, STARTUP_MESSAGE
+from .shell_lib import Shell, STARTUP_MESSAGE
