+++ ontospy/extras/sparqlpy.py
